+++ parsers/ocaml.c
@@ -337,11 +337,11 @@ static void readIdentifier (lexingState * st)
 
 	/* first char is a simple letter */
 	if (isAlpha (*st->cp) || *st->cp == '_')
-		vStringPut (st->name, (int) *st->cp);
+		vStringPut (st->name, *st->cp);
 
 	/* Go till you get identifier chars */
 	for (p = st->cp + 1; isIdent (*p); p++)
-		vStringPut (st->name, (int) *p);
+		vStringPut (st->name, *p);
 
 	st->cp = p;
 }
